Change the UI directory structure to follow the atomic design
