+++ packages/ui/src/ui/article/organisms/ArticleCardList.vue
@@ -1,13 +1,9 @@
 <template>
   <div>
     <template v-for="article in articleList.items">
-      <ArticleCardLink
+      <ArticleCard
         :article="article"
-        v-if="!article.deletedAt"
-        @delete="deleteArticle(article.id)"></ArticleCardLink>
-      <ArticleCardLinkDisabled
-        :article="article"
-        v-if="!!article.deletedAt"></ArticleCardLinkDisabled>
+        @delete="deleteArticle(article.id)" />
     </template>
   </div>
 </template>
@@ -16,13 +12,11 @@
   import {Vue, Component} from 'vue-property-decorator';
   import {mapState} from 'vuex';
   import {IRootState} from '@/store';
-  import ArticleListItemCard from '@/ui/article/components/ArticleCard.vue';
   import {deleteArticle} from '@/di/provider';
-  import ArticleCardLink from '@/ui/article/components/ArticleCardLink.vue';
-  import ArticleCardLinkDisabled from '@/ui/article/components/ArticleCardLinkDisabled.vue';
+  import ArticleCard from '@/ui/article/organisms/ArticleCard.vue';
 
   @Component({
-    components: {ArticleCardLinkDisabled, ArticleCardLink, ArticleListItem: ArticleListItemCard},
+    components: {ArticleCard},
     computed: {
       ...mapState({
         articleList: (state: IRootState) => state.articleListModule.articleList,
